Ensure that we don't add duplicate Swift package references
The PBXProject methods are not idempotent, and it does not
expose the functionality necessary to remove them normally,
so we have to use reflection to safely add the package.
Otherwise, an additional reference is added to the Xcode project
each time the project is built in Unity, causing Xcode errors.