Fixed 2 warnings coming from Embrace Package
Embrace_Stub.cs was utilizing the `?` operator without using #nullable. Certain versions of C# require this.

EmbraceStartupArgs was overriding Equals without also overriding GetHashCode(). This is important if you need to get a hashcode using the data members. This also means the data members need to be readonly, but that is a good thing because it will force a new startup arg to be created instead of updating data on an existing one which could cause issues