Workaround bug in Linux Android Support package (#137)
* Workaround bug in Linux Android Support package

This package does not seem to get installed properly
by the Unity Hub. It's left in a partially installed
state, with a TargetSupport.pkg.tmp folder with a
Payload file inside. This adds an additional Makefile
command to extract that Payload file.

* Skip Android module extraction for non-Android builds

Added a conditional to the Makefile to skip manual extraction of the Android module when the build target is not Android, preventing unnecessary operations for other platforms.

* Update Makefile

* Update ProjectVersion.txt

* Update Unity editor and package versions

Bumped Unity editor to 2022.3.62f3 and updated several package dependencies in manifest.json to their latest versions for improved stability and features.

---------

Co-authored-by: Benjamin Bennett <benjabenn@gmail.com>